+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.codingbuffalo.aerialdream">
+    package="com.sailu.aliendream">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -23,7 +23,7 @@
         android:usesCleartextTraffic="true">
 
         <activity
-            android:name=".SettingsActivity"
+            android:name="com.sailu.aliendream.SettingsActivity"
             android:banner="@drawable/ic_banner"
             android:configChanges="orientation|keyboardHidden|screenSize"
             android:theme="@style/SettingsTheme">
@@ -37,12 +37,12 @@
         </activity>
 
         <activity
-            android:name=".TestActivity"
+            android:name="com.sailu.aliendream.TestActivity"
             android:configChanges="orientation|keyboardHidden|screenSize"
             android:exported="true" />
 
         <service
-            android:name=".AerialDream"
+            android:name="com.sailu.aliendream.AlienDream"
             android:exported="true"
             android:label="@string/daydream_name"
             android:permission="android.permission.BIND_DREAM_SERVICE">
