+++ app/src/main/java/com/sailu/aliendream/data/Apple2018Video.java
@@ -1,4 +1,4 @@
-package com.codingbuffalo.aerialdream.data;
+package com.sailu.aliendream.data;
 
 public class Apple2018Video extends AppleNewVideo {
 }
