+++ app/src/main/java/com/sailu/aliendream/data/Apple2019Video.java
@@ -1,4 +1,4 @@
-package com.codingbuffalo.aerialdream.data;
+package com.sailu.aliendream.data;
 
 public class Apple2019Video extends AppleNewVideo {
 }
