+++ app/src/main/java/com/sailu/aliendream/data/protium/Interactor.java
@@ -1,4 +1,4 @@
-package com.codingbuffalo.aerialdream.data.protium;
+package com.sailu.aliendream.data.protium;
 
 import androidx.annotation.NonNull;
