+++ app/src/main/java/com/sailu/aliendream/data/protium/SimpleTask.java
@@ -1,4 +1,4 @@
-package com.codingbuffalo.aerialdream.data.protium;
+package com.sailu.aliendream.data.protium;
 
 public abstract class SimpleTask extends Task<Void> {
     @Override
