+++ app/src/main/res/layout/settings_activity.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/fragment"
-    android:name="com.codingbuffalo.aerialdream.SettingsFragment"
+    android:name="com.sailu.aliendream.SettingsFragment"
     android:layout_width="match_parent"
     android:layout_height="wrap_content" />
