+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
-    <string name="daydream_name">Aerial Dream</string>
-    <string name="loading_aerial">Loading Aerial…</string>
+    <string name="daydream_name">Alien Dream</string>
+    <string name="loading_aerial">Loading Dream…</string>
 
     <string name="category_settings">Settings</string>
